refactor: 카운터 폼에 적절한 네이밍 부여
